[client,android] Flatten bookmark hierarchy
